preliminary code for happiest state
this code is supposed to pull in tweets and then tag them according to location and mood of the tweet, then find the state which has the best mood.  it does not work.  right now i am having trouble using the 'location' key.  it is embedded in the 'user' key so i am not sure how to work with nested keys.  I had a similar problem in the hashtag program.